#154: Move session_manager to common.database
- This is code for the database backend, so it seems fitting to move it there
- This commit also includes the changes in imports and updates to documentation
