+++ common/database/session_manager.py
